Move some assertion methods from ReleaseTransactionBuilderTest to ReleaseTransactionAssertions
