+++ rskj-core/src/test/java/co/rsk/peg/ReleaseTransactionAssertions.java
@@ -1,14 +1,25 @@
 package co.rsk.peg;
 
-import co.rsk.bitcoinj.core.*;
-import co.rsk.bitcoinj.script.Script;
-
-import java.util.List;
-import java.util.function.Predicate;
-
 import static co.rsk.peg.bitcoin.BitcoinTestAssertions.assertP2shP2wshWitnessWithoutSignaturesHasProperFormat;
+import static co.rsk.peg.bitcoin.BitcoinTestAssertions.assertScriptSigFromP2shErpWithoutSignaturesHasProperFormat;
+import static co.rsk.peg.bitcoin.BitcoinTestAssertions.assertScriptSigFromStandardMultisigWithoutSignaturesHasProperFormat;
+import static co.rsk.peg.bitcoin.BitcoinTestUtils.MIN_NON_DUST_VALUE_FOR_P2SH_OUTPUT_SCRIPT;
+import static co.rsk.peg.bitcoin.BitcoinUtils.BTC_TX_VERSION_1;
 import static co.rsk.peg.bitcoin.BitcoinUtils.BTC_TX_VERSION_2;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import co.rsk.bitcoinj.core.Address;
+import co.rsk.bitcoinj.core.BtcTransaction;
+import co.rsk.bitcoinj.core.Coin;
+import co.rsk.bitcoinj.core.NetworkParameters;
+import co.rsk.bitcoinj.core.TransactionInput;
+import co.rsk.bitcoinj.core.TransactionOutput;
+import co.rsk.bitcoinj.core.TransactionWitness;
+import co.rsk.bitcoinj.core.UTXO;
+import co.rsk.bitcoinj.script.Script;
+import java.util.List;
+import java.util.function.Predicate;
 
 public class ReleaseTransactionAssertions {
 
@@ -55,8 +66,59 @@ public class ReleaseTransactionAssertions {
         assertEquals(expectedNumberOfUtxos, matchingUtxos);
     }
 
+    public static void assertBtcTxVersionIs1(BtcTransaction releaseTransaction) {
+        assertEquals(BTC_TX_VERSION_1, releaseTransaction.getVersion());
+    }
+
     public static void assertBtcTxVersionIs2(BtcTransaction btcTransaction) {
         assertEquals(BTC_TX_VERSION_2, btcTransaction.getVersion());
+    }
+
+    public static void assertOutputsWithDustChange(BtcTransaction releaseTransaction,
+        List<TransactionOutput> releaseTransactionChangeOutputs,
+        Coin requestedAmount) {
+        Coin inputTotalAmount = releaseTransaction.getInputSum();
+        Coin originalChangeAmount = inputTotalAmount.subtract(requestedAmount);
+        assertTrue(isDust(originalChangeAmount));
+
+        Coin amountToGetNonDustValue = MIN_NON_DUST_VALUE_FOR_P2SH_OUTPUT_SCRIPT.subtract(originalChangeAmount);
+        Coin amountToSend = requestedAmount.subtract(amountToGetNonDustValue);
+
+        assertOutputsUserAndChangeValues(releaseTransaction, releaseTransactionChangeOutputs, amountToSend, MIN_NON_DUST_VALUE_FOR_P2SH_OUTPUT_SCRIPT);
+    }
+
+    public static void assertOutputsWithNonDustChange(BtcTransaction releaseTransaction,
+        List<TransactionOutput> releaseTransactionChangeOutputs,
+        Coin requestedAmount) {
+        Coin inputTotalAmount = releaseTransaction.getInputSum();
+        Coin expectedChangeAmount = inputTotalAmount.subtract(requestedAmount);
+        assertOutputsUserAndChangeValues(releaseTransaction, releaseTransactionChangeOutputs, requestedAmount, expectedChangeAmount);
+    }
+
+    private static void assertOutputsUserAndChangeValues(BtcTransaction releaseTransaction,
+        List<TransactionOutput> releaseTransactionChangeOutputs,
+        Coin amountToSend,
+        Coin expectedChangeAmount) {
+        Coin changeOutputsAmount = getOutputsAmount(releaseTransactionChangeOutputs);
+        assertEquals(expectedChangeAmount, changeOutputsAmount);
+
+        Coin userOutputsAmount = releaseTransaction.getOutputSum().subtract(changeOutputsAmount);
+        Coin releaseTransactionFees = releaseTransaction.getFee();
+        Coin userOutputsAndFeesAmount = releaseTransactionFees.add(userOutputsAmount);
+        assertEquals(amountToSend, userOutputsAndFeesAmount);
+        Coin inputTotalAmount = releaseTransaction.getInputSum();
+        assertEquals(inputTotalAmount, userOutputsAndFeesAmount.add(changeOutputsAmount));
+    }
+
+    private static Coin getOutputsAmount(List<TransactionOutput> outputs) {
+        return outputs.stream()
+            .map(TransactionOutput::getValue)
+            .reduce(Coin::add)
+            .orElse(Coin.ZERO);
+    }
+
+    private static boolean isDust(Coin expectedChangeAmount) {
+        return expectedChangeAmount.compareTo(MIN_NON_DUST_VALUE_FOR_P2SH_OUTPUT_SCRIPT) < 0;
     }
 
     public static void assertSelectedUtxosBelongToTheInputs(List<UTXO> selectedUtxos, List<TransactionInput> inputs) {
@@ -68,6 +130,33 @@ public class ReleaseTransactionAssertions {
                 .count();
             assertEquals(1, matchingInputs);
         }
+    }
+
+    /** Input count, script format, and selected-UTXO alignment for peg-out / batched flows. */
+    public static void assertReleaseTxInputsStandardMultisig(
+        BtcTransaction tx,
+        int expectedInputCount,
+        Script federationRedeemScript,
+        List<UTXO> federationUtxos,
+        List<UTXO> selectedUtxos) {
+        List<TransactionInput> inputs = tx.getInputs();
+        assertEquals(expectedInputCount, inputs.size());
+        assertReleaseTxInputsHasProperFormatAndBelongsToStandardMultisigFederation(
+            tx, federationRedeemScript, federationUtxos);
+        assertSelectedUtxosBelongToTheInputs(selectedUtxos, inputs);
+    }
+
+    public static void assertReleaseTxInputsP2shErp(
+        BtcTransaction tx,
+        int expectedInputCount,
+        Script federationRedeemScript,
+        List<UTXO> federationUtxos,
+        List<UTXO> selectedUtxos) {
+        List<TransactionInput> inputs = tx.getInputs();
+        assertEquals(expectedInputCount, inputs.size());
+        assertReleaseTxInputsHasProperFormatAndBelongsToP2shErpFederation(
+            tx, federationRedeemScript, federationUtxos);
+        assertSelectedUtxosBelongToTheInputs(selectedUtxos, inputs);
     }
 
     public static void assertReleaseTxInputsP2shP2wshErp(
@@ -85,6 +174,28 @@ public class ReleaseTransactionAssertions {
         assertSelectedUtxosBelongToTheInputs(selectedUtxos, inputs);
     }
 
+    public static void assertReleaseTxInputsHasProperFormatAndBelongsToStandardMultisigFederation(
+        BtcTransaction releaseTransaction,
+        Script federationRedeemScript,
+        List<UTXO> federationUTXOs) {
+        for (TransactionInput input : releaseTransaction.getInputs()) {
+            Script scriptSig = input.getScriptSig();
+            assertScriptSigFromStandardMultisigWithoutSignaturesHasProperFormat(scriptSig, federationRedeemScript);
+            assertInputIsFromFederationUTXOsWallet(input, federationUTXOs);
+        }
+    }
+
+    public static void assertReleaseTxInputsHasProperFormatAndBelongsToP2shErpFederation(
+        BtcTransaction releaseTransaction,
+        Script federationRedeemScript,
+        List<UTXO> federationUTXOs) {
+        for (TransactionInput input : releaseTransaction.getInputs()) {
+            Script scriptSig = input.getScriptSig();
+            assertScriptSigFromP2shErpWithoutSignaturesHasProperFormat(scriptSig, federationRedeemScript);
+            assertInputIsFromFederationUTXOsWallet(input, federationUTXOs);
+        }
+    }
+
     private static void assertReleaseTxInputsHasProperFormatAndBelongsToP2shP2wshErpFederation(
         BtcTransaction releaseTransaction,
         Script federationRedeemScript,
@@ -96,5 +207,58 @@ public class ReleaseTransactionAssertions {
             assertP2shP2wshWitnessWithoutSignaturesHasProperFormat(witness, federationRedeemScript);
             assertInputIsFromFederationUTXOsWallet(inputs.get(inputIndex), federationUtxos);
         }
+    }
+
+    /**
+     * Like {@link #assertReleaseTxInputsStandardMultisig} but for migration: all retiring federation UTXOs
+     * are spent and {@code selectedUtxos} must match that set exactly.
+     */
+    public static void assertMigrationReleaseTxInputsStandardMultisig(
+        BtcTransaction migrationTransaction,
+        Script retiringFederationRedeemScript,
+        List<UTXO> retiringFederationUtxos,
+        List<UTXO> selectedUtxos) {
+        List<TransactionInput> inputs = migrationTransaction.getInputs();
+        assertEquals(retiringFederationUtxos.size(), inputs.size());
+        assertReleaseTxInputsHasProperFormatAndBelongsToStandardMultisigFederation(
+            migrationTransaction, retiringFederationRedeemScript, retiringFederationUtxos);
+        assertEquals(retiringFederationUtxos, selectedUtxos);
+        assertSelectedUtxosBelongToTheInputs(selectedUtxos, inputs);
+    }
+
+    public static void assertMigrationReleaseTxInputsP2shErp(
+        BtcTransaction migrationTransaction,
+        Script retiringFederationRedeemScript,
+        List<UTXO> retiringFederationUtxos,
+        List<UTXO> selectedUtxos) {
+        List<TransactionInput> inputs = migrationTransaction.getInputs();
+        assertEquals(retiringFederationUtxos.size(), inputs.size());
+        assertReleaseTxInputsHasProperFormatAndBelongsToP2shErpFederation(
+            migrationTransaction, retiringFederationRedeemScript, retiringFederationUtxos);
+        assertEquals(retiringFederationUtxos, selectedUtxos);
+        assertSelectedUtxosBelongToTheInputs(selectedUtxos, inputs);
+    }
+
+    public static void assertMigrationReleaseTxInputsP2shP2wshErp(
+        BtcTransaction migrationTransaction,
+        Script retiringFederationRedeemScript,
+        List<UTXO> retiringFederationUtxos,
+        List<UTXO> selectedUtxos) {
+        assertReleaseTxInputsP2shP2wshErp(
+            migrationTransaction,
+            retiringFederationRedeemScript,
+            retiringFederationUtxos,
+            selectedUtxos,
+            retiringFederationUtxos.size()
+        );
+        assertEquals(retiringFederationUtxos, selectedUtxos);
+    }
+
+    public static void assertReleaseTxNumberOfOutputs(
+        int expectedNumberOfOutputs,
+        List<TransactionOutput> releaseTransactionOutputs
+    ) {
+        int actualNumberOfOutputs = releaseTransactionOutputs.size();
+        assertEquals(expectedNumberOfOutputs, actualNumberOfOutputs);
     }
 }
